+++ test/entrypoint/srv_vshard_custom.lua
@@ -24,7 +24,6 @@ package.preload['customers-storage'] = function()
                     },
                     if_not_exists = true,
                     engine = engine,
-                    id = 542,
                 })
 
                 customers_space:create_index('pk', {